fix(apps/web): ImageManager 단일로 관리
