--- apps/web/src/components/common/ImageManager/TypeB/TypeB.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export function TypeB() {
-  return <div>TypeB</div>;
-}
