+++ apps/web/src/components/common/ImageManager/types.ts
