+++ apps/web/src/components/common/index.ts
@@ -1,5 +1,5 @@
 export { ImageManager } from './ImageManager/ImageManager';
-export type { ImageManagerTypeAProps } from './ImageManager/ImageManager';
+export type { ImageManagerProps } from './ImageManager/ImageManager';
 
 export { MainBreadcrumbItem } from './MainBreadcrumbItem/MainBreadcrumbItem';
